# docs/openapi.yml
openapi: 3.0.3
info:
  title: Tournament Platform API
  version: v1
  description: |
    API for the Tournament Platform, enabling user management, wallet operations, tournament participation, and administrative functions.
    Key features include secure authentication, real-time match updates, and role-based access control.
  contact:
    name: Support Team
    email: support@example-tournament.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:{port}/api/v1 # Define port in variables or replace directly
    description: Local development server
    variables:
      port:
        default: '3000'
        description: Port for the local server
  - url: https://staging.example-tournament.com/api/v1
    description: Staging server
  - url: https://api.example-tournament.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access Token for authentication. Example: 'Bearer {token}'"
    refreshTokenCookie: # For documenting the refresh token mechanism
      type: apiKey
      in: cookie
      name: jid # Matches JWT_REFRESH_COOKIE_NAME from .env.example
      description: "HttpOnly cookie containing the Refresh Token."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Validation Error"
        errors: # Optional, for detailed validation errors
          type: array
          items:
            type: string
          example: ["'email' is required"]
        stack: # Optional, for development environment
          type: string

    # Auth Schemas
...
    WalletInitializeDepositRequest:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 20.50
        currency:
          type: string
          maxLength: 3
          example: "USD"
        # Idempotency Key is sent as a header
    WalletInitializeDepositResponse:
      type: object
      properties:
        message:
          type: string
          example: "Deposit initialized. Proceed to payment gateway."
        paymentGatewayUrl: # Or payment details
          type: string
          format: url
          example: "https://payment.gateway.com/pay/xyz123"
        transactionId:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-8901-2345-678901bcdef0"
...
    TransactionHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: "DEPOSIT"
        amount:
          type: number
          format: float
        status:
          type: string
          example: "COMPLETED"
        description:
          type: string
          nullable: true
        transactionDate:
          type: string
          format: date-time
    PaginatedTransactionHistory:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 48
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionHistoryItem'
    WithdrawalRequest:
      type: object
      required:
        - amount
        - currency
        - withdrawalMethodDetails # E.g., bank account, crypto address
      properties:
        amount:
          type: number
          format: float
          minimum: 1.00 # Example minimum withdrawal
          example: 50.00
        currency:
          type: string
          maxLength: 3
          example: "USD"
        withdrawalMethodDetails:
          type: object # Structure depends on payment methods supported
          example: { "type": "paypal", "email": "user@example.com" }
    WithdrawalResponse:
      type: object
      properties:
        message:
          type: string
          example: "Withdrawal request submitted for approval."
        transactionId:
          type: string
          format: uuid

    # Tournament Schemas
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    IdempotencyKeyHeader:
      name: X-Idempotency-Key # Matches IDEMPOTENCY_KEY_HEADER from .env.example
      in: header
      description: Unique key to prevent duplicate processing of requests (e.g., for deposits).
      required: true # For endpoints that need it
      schema:
        type: string
        format: uuid # Recommended format
...
paths:
...
  /wallet/deposit/initialize:
    post:
      tags: [Wallet]
      summary: Initialize the wallet top-up process.
      description: Requires user authentication. An Idempotency Key header is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletInitializeDepositRequest'
      responses:
        '200':
          description: Deposit initialization successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletInitializeDepositResponse'
        '400':
          description: Validation error or idempotency key reuse with different payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
        '409': # Conflict for idempotency
          description: Idempotency key already processed with the same request.
          content:
            application/json: # Should return the original successful response
              schema:
                $ref: '#/components/schemas/WalletInitializeDepositResponse'
        '500':
          description: Internal server error.
...
  /wallet/history:
    get:
      tags: [Wallet]
      summary: Get transaction history for the logged-in user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [DEPOSIT, WITHDRAWAL, TOURNAMENT_FEE, PRIZE_PAYOUT, REFUND]
          description: Filter by transaction type.
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, COMPLETED, FAILED, CANCELED, REQUIRES_APPROVAL]
          description: Filter by transaction status.
      responses:
        '200':
          description: A paginated list of transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionHistory'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
  /wallet/withdrawals:
    post:
      tags: [Wallet]
      summary: Submit a withdrawal request for finance manager approval.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader' # Good practice for withdrawals too
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
      responses:
        '202': # Accepted for processing
          description: Withdrawal request submitted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          description: Validation error (e.g., insufficient balance, invalid amount).
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
...